# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.GoogleLoginRequest:
    properties:
      access_token:
        type: string
    type: object
  handlers.LogoutRequest:
    properties:
      refresh:
        type: string
    type: object
  handlers.RefreshTokenRequest:
    properties:
      refresh:
        type: string
    type: object
...
  util.HttpError:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  util.TokenPair:
    properties:
      access:
        type: string
      refresh:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive API for generating and managing stories with AI
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Story API
  version: "1.0"
paths:
  /auth/google:
    post:
      consumes:
      - application/json
      description: Accepts a Google access token, validates it, and returns a local
        JWT pair (access and refresh tokens). If the user doesn't exist, a new user
        profile is created.
      parameters:
      - description: Google Access Token
        in: body
        name: google_login_request
        required: true
        schema:
          $ref: '#/definitions/handlers.GoogleLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TokenPair'
        "400":
          description: Invalid request body or missing access_token
          schema:
            $ref: '#/definitions/util.HttpError'
        "401":
          description: Invalid Google token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      summary: Google Login
      tags:
      - Authentication
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user by invalidating their refresh token. This is a
        client-side operation that clears tokens from storage.
      parameters:
      - description: Refresh Token to Invalidate
        in: body
        name: logout_request
        required: true
        schema:
          $ref: '#/definitions/handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or missing refresh token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      summary: User Logout
      tags:
      - Authentication
  /auth/token/refresh:
    post:
      consumes:
      - application/json
      description: Accepts a refresh token and returns a new, short-lived access token.
      parameters:
      - description: Refresh Token
        in: body
        name: refresh_token_request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"access\":\"new_access_token\", \"refresh\":\"original_refresh_token\"}'
          schema:
            $ref: '#/definitions/util.TokenPair'
        "400":
          description: Invalid request body or missing refresh token
          schema:
            $ref: '#/definitions/util.HttpError'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Failed to generate new access token
          schema:
            $ref: '#/definitions/util.HttpError'
      summary: Refresh Access Token
      tags:
      - Authentication
  /auth/w/logout:
    post:
      consumes:
      - application/json
      description: Logs out a user by invalidating their refresh token. This is a
        client-side operation that clears tokens from storage.
      parameters:
      - description: Refresh Token to Invalidate
        in: body
        name: logout_request
        required: true
        schema:
          $ref: '#/definitions/handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or missing refresh token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/util.HttpError'
      summary: User Logout
      tags:
      - Authentication
  /auth/w/token/refresh:
    post:
      consumes:
      - application/json
      description: Accepts a refresh token and returns a new, short-lived access token.
      parameters:
      - description: Refresh Token
        in: body
        name: refresh_token_request
        required: true
        schema:
          $ref: '#/definitions/handlers.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '{\"access\":\"new_access_token\", \"refresh\":\"original_refresh_token\"}'
          schema:
            $ref: '#/definitions/util.TokenPair'
        "400":
          description: Invalid request body or missing refresh token
          schema:
            $ref: '#/definitions/util.HttpError'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/util.HttpError'
        "500":
          description: Failed to generate new access token
          schema:
            $ref: '#/definitions/util.HttpError'
      summary: Refresh Access Token
      tags:
      - Authentication
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"